// packages/core/.stryker-tmp/sandbox-zGjffw/src/types.ts
export interface ImageLabelPluginOptions {
  verticalAlign?: 'top' | 'middle' | 'bottom';
  horizontalAlign?: 'start' | 'middle' | 'end';
  offset?: number;
  imagesList: Array<{
    imageUrl: string;
    imageWidth: number;
    imageHeight: number;
  }>;
}